RN-195: Code cleanup and restructure
